^ fix task

+++ app/Http/Controllers/TaskController.php
@@ -25,6 +25,8 @@ class TaskController extends Controller
             }
             return $next($request);
         });
+
+        $this->bucket_folder = 'files';
     }
 
     /**
@@ -35,7 +37,7 @@ class TaskController extends Controller
     public function index()
     {
         try {
-            $tasks = Task::with('responsible_person','files')->orderBy('updated_at', 'desc')->get();
+            $tasks = Task::with('responsible_person','files')->orderBy('created_at', 'desc')->get();
             $ret['tasks'] = $tasks;
             $ret['user'] = Auth::user();
     
@@ -85,13 +87,16 @@ class TaskController extends Controller
             ]);
     
             if ($validator->fails()) {
-                return $validator->errors();
+                return redirect()->back()->withErrors($validator->errors());
             }
+
+            $ON_PROGRESS = 1;
 
             $task = new Task;
             $task->created_by = Auth::id();
             $task->name = $request->name;
             $task->description = $request->description ?? '';
+            $task->status = $ON_PROGRESS;
             $task->is_history_file_active = (int)$request->is_history_active;
             $task->assign_to = $request->responsible_person == null ? json_encode([]) : json_encode($request->responsible_person);
             $task->save();
@@ -100,7 +105,7 @@ class TaskController extends Controller
             $task->responsible_person()->attach($responsible_ids);
 
             if ($request->hasFile('default_file')) {
-                $this->upload_to_s3($task, $request->file('default_file'), $request->custom_name);
+                $this->upload_file_default_to_s3($task, $request->file('default_file'), $request->custom_name);
             }
 
             DB::commit();
@@ -128,9 +133,18 @@ class TaskController extends Controller
         } catch (\Exception $e) {
             throw new \Exception($e->getMessage());
         }
-        $task = Task::with('responsible_person','files')->where('id', $decrypted_id)->firstOrFail();
+        $task = Task::with(['responsible_person','files' => function($q) {
+                $q->where('is_default', 0)->with('user', 'status')->orderBy('created_at', 'desc');
+            }])
+            ->where('id', $decrypted_id)->firstOrFail();
+        
+        $default_file = Task::with(['files' => function($q) {
+            $q->where('is_default', 1)->orderBy('created_at', 'desc')->limit(1);
+        }])
+        ->findOrFail($decrypted_id);
+        
         $ret['task'] = $task;
-
+        $ret['default_file'] = $default_file->files ? $default_file->files[0] : null;
         return view('tasks.show', $ret);
         
     }
@@ -170,7 +184,6 @@ class TaskController extends Controller
         try {
             DB::beginTransaction();
             $data = $request->except('_method','_token','submit');
-
             try {
                 $decrypted_id = decrypt($id);
             } catch (\Exception $e) {
@@ -182,13 +195,13 @@ class TaskController extends Controller
                 'description' => 'required|string',
                 'is_history_active' => 'required',
                 'assign_to' => 'nullable|string',
-                'default_file' => 'nullable|mimes:pdf,xlx,csv,doc,docx,ppt,pptx,rtf,txt',
+                'default_file' => 'nullable|mimes:pdf,xls,csv,doc,docx,ppt,pptx,rtf,txt,xlsx',
                 'responsible_person' => 'nullable',
                 'custom_name' => 'nullable|string'
             ]);
 
             if ($validator->fails()) {
-                return $validator->errors();
+                return redirect()->back()->withErrors($validator->errors());
             }
 
             $task = Task::with('files')->findOrFail($decrypted_id);
@@ -198,22 +211,24 @@ class TaskController extends Controller
 
                     // Delete Old File
                     $old_file = $task->files->last();
-                    $old_name_file = $old_file->new_name;
-                    $exists = Storage::disk('s3')->exists('files/'.$old_name_file);
-                    if ($exists) {
-                        Storage::disk('s3')->delete('files/'.$old_name_file);
+                    if ($old_file) {
+                        $old_name_file = $old_file->new_name;
+                        $exists = Storage::disk('s3')->exists('files/'.$old_name_file);
+                        if ($exists) {
+                            Storage::disk('s3')->delete('files/'.$old_name_file);
+                        }
+                        $old_file->delete();
                     }
-                    $old_file->delete();
 
                     // Upload new File
                     if ($request->hasFile('default_file')) {
-                        $this->upload_to_s3($task, $request->file('default_file'), $request->custom_name);
+                        $this->upload_file_default_to_s3($task, $request->file('default_file'), $request->custom_name);
                     }
 
                 } else {
                     // Upload new File
                     if ($request->hasFile('default_file')) {
-                        $this->upload_to_s3($task, $request->file('default_file'), $request->custom_name);
+                        $this->upload_file_default_to_s3($task, $request->file('default_file'), $request->custom_name);
                     }
                 }
             }
@@ -308,13 +323,56 @@ class TaskController extends Controller
         }
     }
 
-    public function send_file_task(Request $request)
+    public function send_file_task(Request $request, $id)
     {
         try {
             $validator = Validator::make($request->all(), [
-                'task_id' => 'required|string',
-                'file' => 'required|string'
+                'custom_name' => 'nullable|string',
+                'description' => 'nullable|string',
+                'task_file' => 'required|mimes:pdf,xlx,csv,doc,docx,ppt,pptx,rtf,txt'
             ]);
+
+            if ($validator->fails()) {
+                return redirect()->back()->withErrors($validator->errors());
+            }
+
+            try {
+                $task_id = decrypt($id);
+            } catch (\Exception $e) {
+                throw new \Exception($e->getMessage());
+            }
+
+            $task = Task::where('id', $task_id)->firstOrFail();
+
+            $custom_name = $request->custom_name;
+            $description = $request->description;
+
+            if (!$task->is_history_file_active) {
+                // Delete Old File
+                $old_file = $task->files->where('is_default', '!=', 1)->last();
+                if ($old_file) {
+                    $old_name_file = $old_file->new_name;
+                    $exists = Storage::disk('s3')->exists('files/'.$old_name_file);
+                    if ($exists) {
+                        Storage::disk('s3')->delete('files/'.$old_name_file);
+                    }
+                    $old_file->delete();
+                }
+
+                // Upload new File
+                if ($request->hasFile('task_file')) {
+                    $this->upload_to_s3($task, $request->file('task_file'), $custom_name, $description);
+                }
+
+            } else {
+                // Upload new File
+                if ($request->hasFile('task_file')) {
+                    $this->upload_to_s3($task, $request->file('task_file'), $custom_name, $description);
+                }
+            }
+
+            $request->session()->flash('task.file_uploaded', 'Dokumen berhasil diunggah!');
+            return redirect()->route('tasks.show', encrypt($task->id));
 
         } catch (\Exception $e) {
             dd($e);
@@ -357,7 +415,7 @@ class TaskController extends Controller
         $file_upload->save();
     }
 
-    public function upload_to_s3($task, $file, $custom_name)
+    public function upload_file_default_to_s3($task, $file, $custom_name)
     {
         $filename = $file->getClientOriginalName();
         if ($custom_name) {
@@ -382,6 +440,39 @@ class TaskController extends Controller
         $file_upload->created_by = Auth::id();
         $file_upload->original_name = $original_name;
         $file_upload->description = 'Default File';
+        $file_upload->mime_type = $mime_type;
+        $file_upload->new_name = $new_name;
+        $file_upload->path = $path;
+        $file_upload->is_default = $DEFAULT_FILE;
+        $file_upload->type = $TYPE_FILE;
+        $file_upload->save();
+    }
+
+    public function upload_to_s3($task, $file, $custom_name, $description)
+    {
+        $filename = $file->getClientOriginalName();
+        if ($custom_name) {
+            $original_name = $custom_name;
+        } else {
+            $original_name = pathinfo($filename, PATHINFO_FILENAME);
+        }
+
+        $mime_type = $file->getMimeType();
+        $extension = $file->getClientOriginalExtension();
+
+        $path = Storage::disk('s3')->put($this->bucket_folder, $file);
+        $new_name = basename($path);
+
+        $STATUS_APPROVAL = 2; // waiting aproval
+        $DEFAULT_FILE = 0;
+        $TYPE_FILE = 'internal';
+
+        $file_upload = new File;
+        $file_upload->task_id = $task->id;
+        $file_upload->status_approve = $STATUS_APPROVAL;
+        $file_upload->created_by = Auth::id();
+        $file_upload->original_name = $original_name;
+        $file_upload->description = $description;
         $file_upload->mime_type = $mime_type;
         $file_upload->new_name = $new_name;
         $file_upload->path = $path;
